fix: submit redpacket
